Added new attack hit sprite, base for state manager and other small value tweaks
Created a new sprite for when the player hits a basic attack. Finally finished the base for the Logic State Machine Component but still need to move action functionality from the player character to the Base State and create derived states for each action. Made some small tweaks such as launching the player upwards when hitting an attack while mid-air and adding more safety checks in code for pointers.