Finished logic state machine and states: Idle, Moving, InAir, and Pierce
Implemented the logic state machine but may need to make a few tweaks and go through it all to perhaps find small optimisations. I now have a base for other states such as wall jumping and I can use this as a sort of template for enemy states as well if need be.