Added wall jump state and implementation, and might have fixed errors when accessing objects
Now that the logic state machine has been implemented, it was quite easy to implement wall jumping that is only accessible while the player is in the air. While the player is in the air, if they collide with a wall while holding a movement key in the direction of it, they will begin sliding down it, lowering their Z velocity and allowing them to launch off the wall by jumping, where they are also given an additional jump afterwards. I have also added UPROPERTY() tags to some objects used by the player in the hopes it will fix some errors causing crashing where the objects are sometimes not accessible.