Fixed wall jumping issue, edge sliding and removed unnecessary player character components
While thinking of my wall jump implementation, I figured I'd need 2 colliders on each side of the player to handle wall jumping on each side however during implementation, I figured that I only need the one collider that is positioned in front of the player since. This is because collision only works if the player is moving into the wall while mid-air, and because theyre moving towards the wall, the actor is rotated to face that direction, causing the player's wall collider component to face the wall regardless.

While implementing coyote time, I also found an issue when walking of ledges where, because the player's main collider is a capsule with rounded edges, when running off a ledge, the player would slide down a little bit and when moving slowly, would just slide of the corner edge. To fix this, I simply ticked the "Use flat base for floor checks" in CharacterMovement.

Due to me increasing the width of the capsule component, wall jumping had stopped working properly. To fix this I simply widened the wall collider and moved it a little further away from the player so their capsule component was not blocking before the wall collider could overlap.