Fixed OnWall and jump cutting bugs, implemented basic thrust ability
Main thing is I implemented the players "thrust" ability which acts a less constrained "pierce". This ability will most likely have a cooldown added to it that will be reset by the players pierce. I am still deciding if having both of these is necessary though so one may be removed in the future.

Also fixed a couple bugs: one being that when in the OnWall state, moving in the opposite direction of the wall was putting the player in the "Move" state which caused the coyote time to kick in, teleporting the player downwards. This was fixed by simply transferring them to the "InAir" state instead of the "Move" state. The other bug involved the players jump being cut when they release the jump key, even if they had transferred to another state while the key was still held down and released after. The jump cut should only be for that current jumps "InAir" state so I moved the jump cutting code from the PlayerBaseState class (that was a parent of all the states) to just the "InAir" state, and the variable to determine if a jump will be cut is reset on entering any state.