Implemented ranged weapon heat mechanic in blueprint
An idea for the game is to give the players main weapon, their energy rifle, a heat/overheat mechanic where firing projectiles increases the heat of the weapon and once it reaches max, goes into an overheating stage where it is unusable. After not firing for a small delay, the heat is quickly reduced back to its minimum. I think there could be various perks/properties affected by heat that the player can obtain on their runs that affect gameplay.